refa: move ToJson to api-client
